# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "tt-scrap",
...
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/tiktok/extractions": {
      "post": {
        "tags": [
          "tiktok"
        ],
        "summary": "Extract Tiktok",
        "description": "Extract a TikTok post and return short-lived opaque asset references.",
        "operationId": "extractTikTok",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TikTokExtractionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TikTokExtractionResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid tt-scrap bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Request validation or delivery-parameter error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "default": {
            "description": "Stable tt-scrap error envelope; inspect error.code",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v1/tiktok/music": {
      "post": {
        "tags": [
          "tiktok"
        ],
        "summary": "Extract Tiktok Music",
        "description": "Extract the music attached to a TikTok post by numeric post ID.",
        "operationId": "extractTikTokMusic",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TikTokMusicRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TikTokMusicResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid tt-scrap bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Request validation or delivery-parameter error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "default": {
            "description": "Stable tt-scrap error envelope; inspect error.code",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/v1/instagram/extractions": {
      "post": {
        "tags": [
          "instagram"
        ],
        "summary": "Extract Instagram",
        "description": "Extract an Instagram post into ordered opaque media and thumbnail references.",
        "operationId": "extractInstagram",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InstagramExtractionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InstagramExtractionResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid tt-scrap bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Request validation or delivery-parameter error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "default": {
            "description": "Stable tt-scrap error envelope; inspect error.code",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "AssetDescriptor": {
        "properties": {
          "asset_id": {
            "type": "string",
            "title": "Asset Id",
            "description": "Opaque asset token; do not parse or persist it"
          },
          "kind": {
            "type": "string",
            "enum": [
              "video",
              "image",
              "audio",
              "cover",
              "thumbnail"
            ],
            "title": "Kind",
            "description": "Logical purpose of this asset"
          },
          "position": {
            "type": "integer",
            "title": "Position",
            "description": "Zero-based order within the post",
            "default": 0
          },
          "download_url": {
            "type": "string",
            "title": "Download Url",
            "description": "Relative authenticated download path. Resolve against the tt-scrap base URL"
          },
          "filename": {
            "type": "string",
            "title": "Filename",
            "description": "Suggested filename before final format detection"
          },
          "content_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content Type",
            "description": "Declared MIME type when known; trust the download response type"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Expires At",
            "description": "Expiry time for this asset reference"
          }
        },
        "type": "object",
        "required": [
          "asset_id",
          "kind",
          "download_url",
          "filename",
          "expires_at"
        ],
        "title": "AssetDescriptor",
        "description": "Temporary opaque reference to media that tt-scrap can download on demand."
      },
      "ErrorDetail": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code",
            "description": "Stable error identifier suitable for branching"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Human-readable diagnostic; wording may change"
          },
          "request_id": {
            "type": "string",
            "title": "Request Id",
            "description": "Correlation ID to retain for logs and support"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message",
          "request_id"
        ],
        "title": "ErrorDetail",
        "description": "Stable machine-readable error details."
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "description": "Error envelope returned by tt-scrap exception handlers."
      },
      "InstagramExtractionRequest": {
        "properties": {
          "url": {
            "type": "string",
            "maxLength": 2083,
            "minLength": 1,
            "format": "uri",
            "title": "Url",
            "description": "Public Instagram post, reel, TV, or story URL"
          },
          "refresh": {
            "type": "boolean",
            "title": "Refresh",
            "description": "Bypass cached extraction and create a new extraction_id",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "url"
        ],
        "title": "InstagramExtractionRequest",
        "description": "Extract metadata and assets from an Instagram post URL.",
        "examples": [
          {
            "refresh": false,
            "url": "https://www.instagram.com/reel/SHORTCODE/"
          }
        ]
      },
      "InstagramExtractionResponse": {
        "properties": {
          "extraction_id": {
            "type": "string",
            "title": "Extraction Id",
            "description": "Cached extraction reference reusable by Instagram Telegram delivery"
          },
          "platform": {
            "type": "string",
            "const": "instagram",
            "title": "Platform",
            "default": "instagram"
          },
          "source_url": {
            "type": "string",
            "title": "Source Url"
          },
          "content_type": {
            "type": "string",
            "enum": [
              "video",
              "image",
              "carousel"
            ],
            "title": "Content Type",
            "description": "Single media kind or an ordered mixed carousel"
          },
          "media": {
            "items": {
              "$ref": "#/components/schemas/InstagramMediaItem"
            },
            "type": "array",
            "title": "Media",
            "description": "Ordered extracted media items"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Expires At",
            "description": "Expiry time for the associated asset references"
          }
        },
        "type": "object",
        "required": [
          "extraction_id",
          "source_url",
          "content_type",
          "media",
          "expires_at"
        ],
        "title": "InstagramExtractionResponse",
        "description": "Normalized Instagram post information and temporary asset references."
      },
      "InstagramMediaItem": {
        "properties": {
          "position": {
            "type": "integer",
            "title": "Position",
            "description": "Zero-based source carousel order"
          },
          "media_type": {
            "type": "string",
            "enum": [
              "video",
              "image"
            ],
            "title": "Media Type",
            "description": "Telegram media type selected for this item"
          },
          "quality": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quality",
            "description": "Upstream quality label when supplied"
          },
          "asset": {
            "$ref": "#/components/schemas/AssetDescriptor",
            "description": "Temporary image or video asset"
          },
          "thumbnail": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AssetDescriptor"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional video thumbnail prepared for Telegram media mode"
          }
        },
        "type": "object",
        "required": [
          "position",
          "media_type",
          "asset"
        ],
        "title": "InstagramMediaItem",
        "description": "One ordered Instagram media item with an optional thumbnail."
      },
...
      "TikTokExtractionRequest": {
        "properties": {
          "url": {
            "type": "string",
            "maxLength": 2083,
            "minLength": 1,
            "format": "uri",
            "title": "Url",
            "description": "Public TikTok video, slideshow, or short URL"
          },
          "refresh": {
            "type": "boolean",
            "title": "Refresh",
            "description": "Bypass the short metadata cache and create a new extraction",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "url"
        ],
        "title": "TikTokExtractionRequest",
        "description": "Extract metadata and temporary asset references from a TikTok URL.",
        "examples": [
          {
            "refresh": false,
            "url": "https://www.tiktok.com/@creator/video/1234567890123456789"
          }
        ]
      },
      "TikTokExtractionResponse": {
        "properties": {
          "extraction_id": {
            "type": "string",
            "title": "Extraction Id",
            "description": "Short-lived extraction reference reusable by Telegram delivery; default TTL 60s"
          },
          "platform": {
            "type": "string",
            "const": "tiktok",
            "title": "Platform",
            "default": "tiktok"
          },
          "source_id": {
            "type": "string",
            "title": "Source Id"
          },
          "source_url": {
            "type": "string",
            "title": "Source Url"
          },
          "resolved_url": {
            "type": "string",
            "title": "Resolved Url"
          },
          "content_type": {
            "type": "string",
            "enum": [
              "video",
              "slideshow"
            ],
            "title": "Content Type",
            "description": "Determines whether media contains one video or ordered slideshow images"
          },
          "cover": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AssetDescriptor"
              },
              {
                "type": "null"
              }
            ]
          },
          "width": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Width"
          },
          "height": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Height"
          },
          "duration_seconds": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Seconds"
          },
          "likes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Likes"
          },
          "views": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Views"
          },
          "media": {
            "items": {
              "$ref": "#/components/schemas/AssetDescriptor"
            },
            "type": "array",
            "title": "Media",
            "description": "One selected best-quality video or ordered slideshow images"
          },
          "music": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TikTokMusicMetadata"
              },
              {
                "type": "null"
              }
            ]
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Expires At"
          }
        },
        "type": "object",
        "required": [
          "extraction_id",
          "source_id",
          "source_url",
          "resolved_url",
          "content_type",
          "media",
          "expires_at"
        ],
        "title": "TikTokExtractionResponse",
        "description": "Normalized TikTok post information with no upstream CDN URLs exposed."
      },
      "TikTokMusicMetadata": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "author": {
            "type": "string",
            "title": "Author"
          },
          "duration_seconds": {
            "type": "integer",
            "title": "Duration Seconds"
          },
          "cover": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AssetDescriptor"
              },
              {
                "type": "null"
              }
            ]
          },
          "audio": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AssetDescriptor"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "title",
          "author",
          "duration_seconds"
        ],
        "title": "TikTokMusicMetadata",
        "description": "Music information embedded in a TikTok extraction response."
      },
      "TikTokMusicRequest": {
        "properties": {
          "video_id": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Video Id",
            "description": "Numeric TikTok post ID"
          }
        },
        "type": "object",
        "required": [
          "video_id"
        ],
        "title": "TikTokMusicRequest",
        "description": "Extract TikTok music using a numeric post ID.",
        "examples": [
          {
            "video_id": 1234567890123456789
          }
        ]
      },
      "TikTokMusicResponse": {
        "properties": {
          "extraction_id": {
            "type": "string",
            "title": "Extraction Id"
          },
          "platform": {
            "type": "string",
            "const": "tiktok",
            "title": "Platform",
            "default": "tiktok"
          },
          "source_id": {
            "type": "string",
            "title": "Source Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "author": {
            "type": "string",
            "title": "Author"
          },
          "duration_seconds": {
            "type": "integer",
            "title": "Duration Seconds"
          },
          "cover": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AssetDescriptor"
              },
              {
                "type": "null"
              }
            ]
          },
          "audio": {
            "$ref": "#/components/schemas/AssetDescriptor"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Expires At"
          }
        },
        "type": "object",
        "required": [
          "extraction_id",
          "source_id",
          "title",
          "author",
          "duration_seconds",
          "audio",
          "expires_at"
        ],
        "title": "TikTokMusicResponse",
        "description": "TikTok music metadata and temporary audio/cover asset references."
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}